Automatic item height placement
Items now look at a new variable in the placeableobjectSO that determines the height of the object. When attempting to place a desktop item on to something it will look at the height of the object to determine where to place the object. Desk items can also now be placed on the ground